+++ src/components/ui/badge/index.tsx
@@ -26,8 +26,8 @@ export interface BadgeProps
   extends React.HTMLAttributes<HTMLDivElement>,
     VariantProps<typeof badgeVariants> {}
 
-function Badge({ className, variant, ...props }: BadgeProps) {
+const Badge: React.FC<BadgeProps> = ({ className, variant, ...props }) => {
   return <div className={cn(badgeVariants({ variant }), className)} {...props} />;
-}
+};
 
 export { Badge, badgeVariants };
