+++ src/components/ui/button/index.tsx
@@ -36,15 +36,15 @@ export interface ButtonProps
   asChild?: boolean;
 }
 
-const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ className, variant, size, asChild = false, ...props }, ref) => {
-    const Comp = asChild ? Slot : 'button';
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(function Button(
+  { className, variant, size, asChild = false, ...props },
+  ref,
+) {
+  const Component = asChild ? Slot : 'button';
 
-    return (
-      <Comp className={cn(buttonVariants({ variant, size, className }))} ref={ref} {...props} />
-    );
-  },
-);
-Button.displayName = 'Button';
+  return (
+    <Component className={cn(buttonVariants({ variant, size }), className)} ref={ref} {...props} />
+  );
+});
 
 export { Button, buttonVariants };
